PLOTTING: Highcharts support & misc
 - Highcharts support for bar and barh (pyramid type still missing)
 - moving the other hc files to another directory as they are deprecated.
 - erasing matplotlib code from decomposition file. A new plotting object PCA Scree is added.
 - Passing more parameters to layout to avoid plotting specific params.
 - code cleaning and formatting.
