+++ verticapy/plotting/_highcharts_tmp/__init__.py
@@ -1,0 +1,27 @@
+"""
+(c)  Copyright  [2018-2023]  OpenText  or one of its
+affiliates.  Licensed  under  the   Apache  License,
+Version 2.0 (the  "License"); You  may  not use this
+file except in compliance with the License.
+
+You may obtain a copy of the License at:
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless  required  by applicable  law or  agreed to in
+writing, software  distributed  under the  License is
+distributed on an  "AS IS" BASIS,  WITHOUT WARRANTIES
+OR CONDITIONS OF ANY KIND, either express or implied.
+See the  License for the specific  language governing
+permissions and limitations under the License.
+"""
+from verticapy.plotting._highcharts_tmp.bar import bar
+from verticapy.plotting._highcharts_tmp.base import hchart_from_vdf
+from verticapy.plotting._highcharts_tmp.boxplot import boxplot
+from verticapy.plotting._highcharts_tmp.candlestick import candlestick
+from verticapy.plotting._highcharts_tmp.drilldown_chart import drilldown_chart
+from verticapy.plotting._highcharts_tmp.heatmap import heatmap
+from verticapy.plotting._highcharts_tmp.line import line
+from verticapy.plotting._highcharts_tmp.negative_bar import negative_bar
+from verticapy.plotting._highcharts_tmp.pie import pie
+from verticapy.plotting._highcharts_tmp.scatter import scatter
+from verticapy.plotting._highcharts_tmp.spider import spider
