+++ verticapy/plotting/_highcharts_tmp/boxplot.py
