+++ verticapy/plotting/_highcharts_tmp/scatter.py
