+++ verticapy/plotting/_highcharts_tmp/spider.py
@@ -20,7 +20,7 @@ from verticapy._config.colors import get_colors
 from verticapy._utils._sql._sys import _executeSQL
 from verticapy.connection import current_cursor
 
-from verticapy.plotting._highcharts.utils import data_to_columns
+from verticapy.plotting._highcharts_tmp.utils import data_to_columns
 
 
 def spider(
